Sourcing the profile files which Heroku buildpacks set up for us (in /app/.profile.d), so that dependencies downloaded by the buildpack are available on PATH.
The docker run command which starts the app is now a little verbose so maybe should be refactored into a separate script; part of the verbosity is because of guarding against absence of the .profile.d directory.

This is to support use of the Java buildpack but has also been tested working for a simple Ruby web app.