docs: remove outdated global resource setting
This was removed in c9bbddcfcbbcad64dd9453afec77ba614598fc67 as it was more complicated for operators to figure out _exactly_ which combination of limits/reservations would apply to their applications. Unfortunately, the documentation was never updated.

We may re-add this functionality in the future, but for now, cleaning up the docs is the best option.

[ci skip]