docs: swap from freenode to libera.chat
Freenode is basically offline at this point and we've had a libera.chat for a while, so we should point to that.

Also default all real-time communication to Slack.

Finally, remove prominent links to IRC. While libera is still available, the lack of monitoring makes it not a great way to provide support to users.